NMS-14541 Bridge Topology Discovery, Move BridgeTopologyException
Reworked Static methods
Removing for classes and use into service
discovers and service impl
Only partial changes
