+++ features/enlinkd/service/api/src/main/java/org/opennms/netmgt/enlinkd/service/api/Bridge.java
@@ -29,60 +29,10 @@
 package org.opennms.netmgt.enlinkd.service.api;
 
 import java.util.HashSet;
-import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 
-import org.opennms.core.utils.InetAddressUtils;
-import org.opennms.netmgt.enlinkd.model.BridgeElement;
-
 public class Bridge implements Topology {
-
-    public static Set<String> getIdentifier(List<BridgeElement> elems) {
-        Set<String> identifiers = new HashSet<>();
-        for (BridgeElement element: elems) {
-            if (InetAddressUtils.isValidBridgeAddress(element.getBaseBridgeAddress())) {
-                identifiers.add(element.getBaseBridgeAddress());
-            }
-
-        }
-        return identifiers;
-    }
-
-    public static String getDesignated(List<BridgeElement> elems) {
-        for (BridgeElement element: elems) {
-            if (InetAddressUtils.
-                    isValidStpBridgeId(element.getStpDesignatedRoot()) 
-                    && !element.getBaseBridgeAddress().
-                    equals(InetAddressUtils.getBridgeAddressFromStpBridgeId(element.getStpDesignatedRoot()))) {
-                String designated=InetAddressUtils.
-                               getBridgeAddressFromStpBridgeId(element.getStpDesignatedRoot());
-                if (InetAddressUtils.isValidBridgeAddress(designated)) {
-                    return designated;
-                }
-            }
-        }
-        return null;
-    }
-
-    public static Bridge create(BroadcastDomain domain, Integer nodeid) {
-        Bridge bridge = new Bridge(nodeid);
-        domain.getBridges().add(bridge);
-        return bridge;
-    }
-
-    public static void createRootBridge(BroadcastDomain domain, Integer nodeid) {
-        Bridge bridge = new Bridge(nodeid);
-        bridge.setRootBridge();
-        domain.getBridges().add(bridge);
-    }
-    
-    public static Bridge create(BroadcastDomain domain, Integer nodeid, Integer rootport) {
-        Bridge bridge = new Bridge(nodeid);
-        bridge.setRootPort(rootport);
-        domain.getBridges().add(bridge);
-        return bridge;
-    }
 
     private final Integer m_nodeId;
     private Integer m_rootPort;
@@ -90,7 +40,7 @@ public class Bridge implements Topology {
     private Set<String> m_identifiers = new HashSet<>();
     private String m_designated;
 
-    private Bridge(Integer id) {
+    public Bridge(Integer id) {
         super();
         m_nodeId = id;
     }
