+++ features/enlinkd/service/api/src/main/java/org/opennms/netmgt/enlinkd/service/api/BridgeForwardingTable.java
@@ -37,125 +37,16 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import org.opennms.netmgt.enlinkd.service.api.BridgeForwardingTableEntry.BridgeDot1qTpFdbStatus;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 
 public class BridgeForwardingTable implements Topology {
-    
-    private static final Logger LOG = LoggerFactory.getLogger(BridgeForwardingTable.class);
 
-    public static Set<BridgePortWithMacs> getThroughSet(BridgeForwardingTable bridgeFt, Set<BridgePort> excluded) throws BridgeTopologyException {
-
-        for (BridgePort exclude: excluded) {
-            if (exclude.getNodeId().intValue() != bridgeFt.getNodeId().intValue()) {
-                throw new BridgeTopologyException("getThroughSet: node mismatch ["
-                        + bridgeFt.getNodeId() + "]", exclude);
-            }
-        }
-        Set<BridgePortWithMacs> throughSet= new HashSet<>();
-        bridgeFt.getPorttomac().stream().filter(ptm ->!excluded.contains(ptm.getPort())).forEach(throughSet::add);
-        return throughSet;
-    }
-
-    public static BridgeForwardingTable create(Bridge bridge, Set<BridgeForwardingTableEntry> entries) throws BridgeTopologyException {
-        if (bridge == null) {
-            throw new BridgeTopologyException("bridge must not be null");
-        }
-        if (entries == null) {
-            throw new BridgeTopologyException("bridge forwarding table must not be null");
-        }
-        
-        for (BridgeForwardingTableEntry link: entries) {
-            if (link.getNodeId().intValue() != bridge.getNodeId().intValue()) {
-                throw new BridgeTopologyException("create: bridge:["+ bridge.getNodeId()+ "] and forwarding table must have the same nodeid", link);                
-            }
-        }
-        final BridgeForwardingTable bridgeFt = new BridgeForwardingTable(bridge,entries);
-
-        entries.stream().filter(link -> link.getBridgeDot1qTpFdbStatus() 
-                                == BridgeDot1qTpFdbStatus.DOT1D_TP_FDB_STATUS_SELF).
-                                forEach(link -> {
-            bridgeFt.getIdentifiers().add(link.getMacAddress());
-            if (LOG.isDebugEnabled()) {
-                LOG.debug("create: bridge:[{}] adding bid {}",
-                          bridge.getNodeId(),
-                          link.printTopology());
-            }
-        });
-        
-        for (BridgeForwardingTableEntry link: entries) {
-            if (link.getBridgeDot1qTpFdbStatus() 
-                                != BridgeDot1qTpFdbStatus.DOT1D_TP_FDB_STATUS_LEARNED ) {
-                continue;
-            }
-            if (bridgeFt.getIdentifiers().contains(link.getMacAddress())) {
-                if (LOG.isDebugEnabled()) {
-                    LOG.debug("create: bridge:[{}] skip bid {}",
-                          bridge.getNodeId(),
-                          link.printTopology());
-                }
-                continue;
-            }
-                
-            BridgePort bridgeport = BridgePort.getFromBridgeForwardingTableEntry(link);
-                
-            BridgePortWithMacs bpwm = bridgeFt.getBridgePortWithMacs(bridgeport);
-            if (bpwm == null ) {
-                bridgeFt.getPorttomac().add(BridgePortWithMacs.create(bridgeport, new HashSet<>()));
-            }
-            bridgeFt.getBridgePortWithMacs(bridgeport).getMacs().add(link.getMacAddress());
-
-            if (bridgeFt.getMactoport().containsKey(link.getMacAddress())) {
-                bridgeFt.getDuplicated().put(link.getMacAddress(), new HashSet<>());
-                bridgeFt.getDuplicated().get(link.getMacAddress()).add(bridgeport);
-
-                if (LOG.isDebugEnabled()) {
-                    LOG.debug("create: bridge:[{}] duplicated {}",
-                              bridge.getNodeId(),
-                              link.printTopology());
-                }
-                continue;
-            }
-            
-            if (LOG.isDebugEnabled()) {
-                    LOG.debug("create: bridge:[{}] adding {}",
-                          bridge.getNodeId(),
-                          link.printTopology());
-            }
-            bridgeFt.getMactoport().put(link.getMacAddress(), bridgeport);
-        }
-
-        for (String mac: bridgeFt.getDuplicated().keySet()) {
-            BridgePort saved = bridgeFt.getMactoport().remove(mac);            
-            if (LOG.isDebugEnabled()) {
-                LOG.debug("create: bridge:[{}] remove duplicated [{}] from {}",
-                          bridge.getNodeId(),
-                          mac,
-                          saved.printTopology());
-            }
-
-            BridgePortWithMacs savedwithmacs = bridgeFt.getBridgePortWithMacs(saved);
-            savedwithmacs.getMacs().remove(mac);
-            
-            for (BridgePort dupli: bridgeFt.getDuplicated().get(mac)) {
-                BridgePortWithMacs dupliwithmacs = bridgeFt.getBridgePortWithMacs(dupli);
-                dupliwithmacs.getMacs().remove(mac);
-            }
-            bridgeFt.getDuplicated().get(mac).add(saved);
-        }
-
-        return bridgeFt;
-    }
-    
     private final Bridge m_bridge;
     private final Set<BridgeForwardingTableEntry> m_entries;
     private Map<String, BridgePort> m_mactoport = new HashMap<>();
     private Map<String, Set<BridgePort>> m_duplicated = new HashMap<>();
     private final Set<BridgePortWithMacs> m_porttomac = new HashSet<>();
 
-    private BridgeForwardingTable(Bridge bridge, Set<BridgeForwardingTableEntry> entries) {
+    public BridgeForwardingTable(Bridge bridge, Set<BridgeForwardingTableEntry> entries) {
         m_bridge = bridge;
         m_entries = entries;
     }
