+++ features/enlinkd/tests/src/test/java/org/opennms/netmgt/enlinkd/EnLinkdIT.java
@@ -62,6 +62,7 @@ import org.opennms.netmgt.enlinkd.model.IpNetToMedia.IpNetToMediaType;
 import org.opennms.netmgt.enlinkd.service.api.BridgeForwardingTableEntry;
 import org.opennms.netmgt.enlinkd.service.api.BridgePort;
 import org.opennms.netmgt.enlinkd.service.api.BridgeTopologyException;
+import org.opennms.netmgt.enlinkd.service.api.BridgeTopologyService;
 import org.opennms.netmgt.enlinkd.service.api.BroadcastDomain;
 import org.opennms.netmgt.enlinkd.service.api.Node;
 import org.opennms.netmgt.enlinkd.service.api.SharedSegment;
@@ -516,7 +517,7 @@ public class EnLinkdIT extends EnLinkdBuilderITCase {
         BroadcastDomain nodeCbd = m_bridgeTopologyService.getBroadcastDomain(nodeC.getId());
         assertEquals(nodeAbd, nodeBbd);
         assertEquals(nodeAbd, nodeCbd);
-        BroadcastDomain.hierarchySetUp(nodeAbd,nodeAbd.getBridge(nodeA.getId()));
+        BridgeTopologyService.hierarchySetUp(nodeAbd,nodeAbd.getBridge(nodeA.getId()));
 
         topology.checkwithshared(nodeAbd);
         assertEquals(0, nodeAbd.getForwarders(topology.nodeAId).size());
@@ -655,7 +656,7 @@ public class EnLinkdIT extends EnLinkdBuilderITCase {
         assertNotNull(nodeAbd.getBridge(nodeA.getId()));
         assertNotNull(nodeAbd.getBridge(nodeB.getId()));
         assertNotNull(nodeAbd.getBridge(nodeC.getId()));
-        BroadcastDomain.hierarchySetUp(nodeAbd,nodeAbd.getBridge(nodeA.getId()));
+        BridgeTopologyService.hierarchySetUp(nodeAbd,nodeAbd.getBridge(nodeA.getId()));
         assertNotNull(nodeAbd.getRootBridge());
         topology.check(nodeAbd);
         
@@ -756,7 +757,7 @@ public class EnLinkdIT extends EnLinkdBuilderITCase {
         assertNotNull(nodeAbd.getBridge(nodeA.getId()));
         assertNotNull(nodeAbd.getBridge(nodeB.getId()));
         assertNotNull(nodeAbd.getBridge(nodeC.getId()));
-        BroadcastDomain.hierarchySetUp(nodeAbd,nodeAbd.getBridge(nodeA.getId()));
+        BridgeTopologyService.hierarchySetUp(nodeAbd,nodeAbd.getBridge(nodeA.getId()));
         assertNotNull(nodeAbd.getRootBridge());
         topology.check(nodeAbd);
 
@@ -859,7 +860,7 @@ public class EnLinkdIT extends EnLinkdBuilderITCase {
         assertNotNull(nodeAbd.getBridge(nodeA.getId()));
         assertNotNull(nodeAbd.getBridge(nodeB.getId()));
         assertNotNull(nodeAbd.getBridge(nodeC.getId()));
-        BroadcastDomain.hierarchySetUp(nodeAbd,nodeAbd.getBridge(nodeA.getId()));
+        BridgeTopologyService.hierarchySetUp(nodeAbd,nodeAbd.getBridge(nodeA.getId()));
         assertNotNull(nodeAbd.getRootBridge());
         topology.check(nodeAbd);
